+++ example/lib/app.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+import 'home.dart';
+class App extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: HomeScreen(),
+    );
+  }
+}