+++ example/lib/home.dart
@@ -1,0 +1,48 @@
+import 'package:flutter/material.dart';
+import 'screens/hello_world.dart';
+import 'screens/custom_object.dart';
+import 'screens/runtime_materials.dart';
+import 'screens/texture_and_rotation.dart';
+
+class HomeScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('ArCore Demo'),
+      ),
+      body: ListView(
+        children: <Widget>[
+          ListTile(
+            onTap: () {
+              Navigator.of(context)
+                  .push(MaterialPageRoute(builder: (context) => HelloWorld()));
+            },
+            title: Text("Hello World"),
+          ),
+          ListTile(
+            onTap: () {
+              Navigator.of(context).push(
+                  MaterialPageRoute(builder: (context) => CustomObject()));
+            },
+            title: Text("Custom Anchored Object with onTap"),
+          ),
+          ListTile(
+            onTap: () {
+              Navigator.of(context).push(
+                  MaterialPageRoute(builder: (context) => RuntimeMaterials()));
+            },
+            title: Text("Change Materials Property in runtime"),
+          ),
+          ListTile(
+            onTap: () {
+              Navigator.of(context).push(MaterialPageRoute(
+                  builder: (context) => ObjectWithTextureAndRotation()));
+            },
+            title: Text("Custom object with texture and rotation listener "),
+          )
+        ],
+      ),
+    );
+  }
+}