+++ lib/src/arcore_android_view.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-/// Callback signature for when a platform view was created.
-///
-/// `id` is the platform view's unique identifier.
 typedef PlatformViewCreatedCallback = void Function(int id);
 
 class ArCoreAndroidView extends AndroidView {