+++ lib/src/shape/arcore_cube.dart
@@ -4,7 +4,6 @@ import 'package:vector_math/vector_math_64.dart';
 
 import 'package:arcore_flutter_plugin/src/shape/arcore_shape.dart';
 
-
 class ArCoreCube extends ArCoreShape {
   ArCoreCube({
     this.size,