+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Add ArCoreNode
+* Add ArCoreRotatingNode
+* Add ArCoreShape (sphere, cube, cylindre)
+* Add shape materials properties (color, texture, metallic, roughness, reflectance)